backend/lambda/detection_pipeline/requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
psycopg2-binary==2.9.9
boto3==1.34.0
numpy==1.24.3
scipy==1.11.4
opencv-python-headless==4.8.1.78
